Add float div by zero test for truediv, make new test file
